api_yamdb/requirements.txt
#asgiref==3.2.10
python-dotenv==0.21.0
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
django_filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.5
pytest-pythonpath==0.7.4
pytest-django==4.5.2
